Generate formatted flakes
Garner now produces flakes with a `formatter` output (currently hardcoded), and
uses `nix fmt` to apply that formatter to the output, avoiding the need to
maintain formatting during generation.

That said, this does reformat the TS flake template to match the hardcoded
formatter (for easier eyeballing), but it also removes some attempts at padding
elements that are spliced into the template.

Fixes #17.