Add a GH workflow for non-flake checks
There are still some checks that can’t yet be done within a flake. This runs
those in a less-constrained GH workflow, so at least CI can verify them before
merging.

This renames `just ci` to `just pre-push` and uses `just ci` for the things that
run in the new workflow.