Test running garner in isolation
This helps ensure that everything necessary to run garner is properly captured.
Spoiler: it’s not.

This also adds a package called “garner” (rather than just “default”)
to make it more obvious when depending on it.

Unfortunately, this is not as strict as it would ideally be. The need for
Internet access makes it infeasible to run this as a Nix check currently.