Update the flake inputs
In the hope that this will solve the current SHA mismatch failures in workflow
jobs.

It also unifies the inputs to reduced the duplicated dependencies.

And `hspec-golden` is pinned to a newer version than in the pinned nixpkgs, but
I think sticking with the release and making that one update is the right thing.