Tighten isolation
The fileserver (and anything other than codegen and garner) is run in the
default shell, with a garner-only shell being used for the parts we're
testing. This means we no longer need `devShells.barren` and the isolation shell
doesn't GHC or git (or potentially anything else) leak into it.

This also allows us to fold this test into `just github-ci`, rather than adding
a new job to the GH workflow.