Don't re-deploy the unversioned ts library
Previously our `release-ts` script would put the temporary directory in
a weird state where `git status` wasn't clean. So it would add the
unversioned library to the created commit, but then not have the files
in it. The `git --work-tree...` did that. It was hard to wrap your head
around. So I basically rewrote that last part. I think this version is
much easier to understand and doesn't put the temporary directory in
that weird state.

I also made the `tag` a just recipe argument, so now you deploy with
e.g.:

```bash
just release-ts 9ee4092 v0.0.11
cd <that temporary directory that it created and printed
git show # or whatever to see the created commit
git push --tags
```

Having the tag as a just argument made it just much easier to work on
this. Otherwise we have to re-type the tag all the time.