Explicit shell/run arguments
Previously we just tacked on $@ to the end of the shell script. But
often this is completely the wrong place for it, and for basically
anything that is a proper shell script (e.g. editGarnConfig) rather than
single command, makes it impossible to correctly pass arguments down.

This PR changes behavior so that you have to explicitly use $@/$*. This
is exactly like any bash script, so I don't think it should be too
confusing.